Rework property loading
Fixes issue #21
